refactor redux to remove cyclic imports
