+++ components/question/QuestionRedux.ts
@@ -1,12 +1,9 @@
 import { connect } from "react-redux";
+import {
+  mapDispatchToProps,
+  mapStateToProps,
+} from "../../store/survey/SurveyReducer";
 
 import Question from "./Question";
-import mapDispatchToProps from "./QuestionD2P";
-
-export function mapStateToProps(state: any, myProps: any) {
-  return {
-    data: state.survey.data,
-  };
-}
 
 export default connect(mapStateToProps, mapDispatchToProps)(Question);
